requirements.txt
# 딥러닝 프레임워크
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# 이미지 처리
Pillow>=10.0.0
opencv-python>=4.8.0

# 데이터 처리
numpy>=1.24.0
pandas>=2.0.0

# 자연어 처리
nltk>=3.8.0
transformers>=4.30.0

# 평가 지표
pycocotools>=2.0.6

# 설정 파일 관리
pyyaml>=6.0

# 진행 상황 표시
tqdm>=4.65.0

# 시각화
matplotlib>=3.7.0
seaborn>=0.12.0

# 유틸리티
tensorboard>=2.13.0

